cnr-ifn.tn/fabrication_facilities/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: BSD License",
]
name = "Fabrication-facilities"
#suggested description, instead of "Plugin for fabrication processes"
description = "Plugin for nanofabrication semiconductor cleanroom processes"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Matteo Bontorno", email = "mbontorno@fbk.eu" },
]
maintainers = [
    { name = "Matteo Bontorno", email = "mbontorno@fbk.eu" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.0",
    "python-magic-bin; sys_platform == 'win32'",
]

[project.urls]
Repository = "https://github.com/foo/Fabrication-facilities"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLC0415",
    "PLR0913",
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
#normalizer_entry_point = "fabrication_facilities.normalizers:normalizer_entry_point"
equipmentapp = "fabrication_facilities.apps:equipment_app_entry_point"
processapp = "fabrication_facilities.apps:process_app_entry_point"
stepapp = "fabrication_facilities.apps:step_app_entry_point"
Items_entry_point = "fabrication_facilities.schema_packages:Items_entry_point"
Utilities_entry_point = "fabrication_facilities.schema_packages:Utilities_entry_point"
Add_entry_point = "fabrication_facilities.schema_packages:Add_entry_point"
Transform_entry_point = "fabrication_facilities.schema_packages:Transform_entry_point"
Remove_entry_point = "fabrication_facilities.schema_packages:Remove_entry_point"
Equipments_entry_point = "fabrication_facilities.schema_packages:Equipments_entry_point"
materials_entry_point = "fabrication_facilities.schema_packages:materials_entry_point"
calculus_entry_point = "fabrication_facilities.schema_packages:calculus_entry_point"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
